+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 Popmotion adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.4.0] 2015-10-16
+
+### Added
+- `Sequence` class, for creating a sequence of non-sequential actions.
+- `onComplete` property to `Iterator.stagger`.
+- Pass an action as the `Iterator.each` and `Iterator.stagger` `method` argument to automatically run `start` on each `Actor`.
+
+### Changed
+- Using rAF-provided timestamp instead of `performance.now` for smoother motion.
+
 ## [3.3.3] 2015-10-13
 
 ### Fixed